ml/model/requirements.txt
torch
torchvision
opencv-python
pyyaml